fix(ci): Verify kubernetes schema against a recent version
The default behaviour of kubeval is to use the most recent kubernetes API available in the https://github.com/instrumenta/kubernetes-json-schema repository.
This repository hasn't been updated since kubernetes v1.18, a version now out of support and not representative of modern kubernetes.

These improvements to the Makefile `validate_chart` target sets a API repo with recent kubernetes api version, and sets the default version to the currently earliest supported version, 1.21.0.